Updating Failing Tests
Tests failed because of bad test data and misused functions in the
RegisteredUser object. Changed code to fix these issues and moved the
RegisteredUserTest to the correct directory.
